unix: Fix external unix inode type

+++ criu/cr-service.c
@@ -304,7 +304,7 @@ static int setup_opts_from_req(int sk, CriuOpts *req)
 	if (req->has_ext_unix_sk) {
 		opts.ext_unix_sk = req->ext_unix_sk;
 		for (i = 0; i < req->n_unix_sk_ino; i++) {
-			if (unix_sk_id_add(req->unix_sk_ino[i]->inode) < 0)
+			if (unix_sk_id_add((unsigned int)req->unix_sk_ino[i]->inode) < 0)
 				goto err;
 		}
 	}
